Assert the zero-allocation claim instead of just stating it
Adds allocation regression tests that run a million codec events and about
an hour of simulated playing, forcing a collection either side and asserting
the retained heap does not grow. Measured growth is ~40 KB where per-event
allocation would show as tens of megabytes.

The claim underpins the whole real-time design and is easy to break with one
innocuous refactor — a .map(), an object literal passed to a callback. This
makes a future change have to defeat it consciously.

Also switches lastStrikeMs to f64: it holds performance.now() values, and
float32 runs out of mantissa past ~16.7 million ms, so a page open for five
hours would start quantising the refractory window. And pins down what
happens when a hovering finger slides into a raised black key — triggering
is correct there, since the finger really is inside the key body, but the
two zones reference different strike planes so it deserves to be intentional.

+++ packages/interaction/src/PokeDetector.ts
@@ -111,8 +111,16 @@ export class PokeDetector {
   private readonly prevDepth = new Float32Array(MAX_FINGERS);
   /** Whether prevDepth holds a usable value. */
   private readonly hasPrev = new Uint8Array(MAX_FINGERS);
-  /** Timestamp of this finger's last note-on, for the refractory window. */
-  private readonly lastStrikeMs = new Float32Array(MAX_FINGERS).fill(-1e9);
+  /**
+   * Timestamp of this finger's last note-on, for the refractory window.
+   *
+   * f64, not f32: these are `performance.now()` values, and float32 runs out of
+   * mantissa past ~16.7 million ms. A page left open for five hours would start
+   * quantising the refractory comparison to several milliseconds, which is the
+   * kind of bug that only appears during a long session and looks like the
+   * hardware misbehaving.
+   */
+  private readonly lastStrikeMs = new Float64Array(MAX_FINGERS).fill(-1e9);
   /** Ring of recent approach speeds (m/s) per finger, for peak detection. */
   private readonly speedHistory = new Float32Array(MAX_FINGERS * VELOCITY_HISTORY);
   private readonly speedCursor = new Uint8Array(MAX_FINGERS);
